+++ Public/get-userperms.ps1
@@ -4,7 +4,7 @@
 # large tenants.
 # Requires: Exchange Online
 
-if (-not (Get-ConnectionInformation)) { Connect-ExchangeOnline -ShowBanner:$false }
+if (-not (Get-ConnectionInformation)) { Connect-ExchangeOnline -ShowBanner:$false -DisableWAM }
 
 $upn = Read-Host "User UPN to check delegated access for"