+++ Public/get-userreport.ps1
@@ -8,7 +8,7 @@
 # Requires: Graph (User.Read.All, Directory.Read.All,
 #           UserAuthenticationMethod.Read.All) + Exchange Online
 
-if (-not (Get-ConnectionInformation)) { Connect-ExchangeOnline }
+if (-not (Get-ConnectionInformation)) { Connect-ExchangeOnline -DisableWAM }
 if (-not (Get-MgContext)) {
     Connect-MgGraph -Scopes "User.Read.All","Directory.Read.All","UserAuthenticationMethod.Read.All" -ContextScope Process
 }